test/requirements.txt
# requirements.txt
pendulum
pymongo
apache-airflow
apache-airflow-providers-mysql
pytest
pytest-cov
certifi
